feat: use single crate for mate development
