+++ src/mate/Cargo.toml
@@ -10,12 +10,27 @@ documentation.workspace = true
 name = "mate"
 path = "src/bin/main.rs"
 
-[dependencies]
-anyhow = { workspace = true }
-bytes = { workspace = true }
-clap = { workspace = true, features = ["env", "derive", "std"] }
-serde = { workspace = true, features = ["derive"] }
-serde_json = { workspace = true}
-tokio = { workspace = true, features = ["macros", "rt-multi-thread"] }
+[lib]
+name = "mate"
+path = "src/lib.rs"
+proc-macro = true
 
-mate_executor = { workspace = true }
+[features]
+default = ["cli"]
+cli = ["dep:anyhow", "dep:bytes", "dep:clap", "dep:serde", "dep:serde_json", "dep:tokio", "dep:mate_executor"]
+macros = ["dep:proc-macro2", "dep:quote", "dep:syn", "dep:wit-bindgen"]
+
+[dependencies]
+anyhow = { workspace = true, optional = true }
+bytes = { workspace = true, optional = true }
+clap = { workspace = true, features = ["env", "derive", "std"], optional = true }
+serde = { workspace = true, features = ["derive"], optional = true }
+serde_json = { workspace = true, optional = true }
+tokio = { workspace = true, features = ["macros", "rt-multi-thread"], optional = true }
+
+proc-macro2 = { workspace = true, optional = true }
+quote = { workspace = true, optional = true }
+syn = { workspace = true, features = ["full", "extra-traits"], optional = true }
+wit-bindgen = { workspace = true, optional = true }
+
+mate_executor = { workspace = true, optional = true }
