--- src/task/Cargo.toml
@@ -1,20 +1,0 @@
-[package]
-name = "mate_task"
-edition.workspace = true
-version.workspace = true
-authors.workspace = true
-description.workspace = true
-documentation.workspace = true
-
-[lib]
-name = "mate_task"
-path = "src/lib.rs"
-proc-macro = true
-
-[dependencies]
-serde = { workspace = true, features = ["derive"] }
-serde_json = { workspace = true }
-proc-macro2 = { workspace = true }
-quote = { workspace = true }
-syn = { workspace = true, features = ["full", "extra-traits"] }
-wit-bindgen = { workspace = true }
