+++ tasks/http/Cargo.toml
@@ -14,4 +14,4 @@ tokio = { workspace = true }
 wit-bindgen = { workspace = true }
 wstd = { workspace = true }
 
-mate_task = { workspace = true }
+mate = { workspace = true, features = ["macros"] }
