+++ tasks/http/src/lib.rs
@@ -1,7 +1,7 @@
 use serde_json::Value;
 use wstd::http::{Body, BodyExt, Client, Method, Request};
 
-use mate_task::{mate_handler, mate_object};
+use mate::{mate_handler, mate_object};
 
 #[mate_object]
 struct Config {
